refactor: configList tags into a component
