+++ src/components/Tags/index.tsx
@@ -1,0 +1,1 @@
+export { Tags } from "./Tags";
